+++ js/functions/formatHashtags.js
@@ -1,3 +1,4 @@
-const formatHashtags = (hashtags, isFormatToLowercase) => isFormatToLowercase ? hashtags.trim().split(/\s+/) : hashtags.trim().toLowerCase().split(/\s+/);
+const formatHashtags = (hashtags, isFormatToLowercase) => isFormatToLowercase ?
+  hashtags.trim().split(/\s+/) : hashtags.trim().toLowerCase().split(/\s+/);
 
 export { formatHashtags };
