+++ js/view/thumbnail.js
